Improve the fit of GLM in deseq2
When the initial fit of LFCs is not stable, the algorithm resorts to a
MLE fit. It turns out that, for our use case, the L-BFSG-B
implementation of scipy is not as effective in Python as it is in R. We
thus resort to a Newton conjugate gradient algorithm that accounts bound
constraints. We also provide a gradient function to the algorithm to
further improve the MLE fit.